% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.background.asc.R
\name{reps.background.asc}
\alias{reps.background.asc}
\title{reps.background.asc, creates Monte Carlo replicates for ENM background test using .asc
input files}
\usage{
reps.background.asc(infile = x, outfile = FALSE, points, reps = 10,
  verbose = FALSE)
}
\arguments{
\item{outfile}{An output .csv file to write replicates to.  If not provided, ENMTools
will return a data frame instead.}

\item{reps}{The number of replicates to create}

\item{verbose}{Controls printing of diagnostic messages}

\item{x}{Path to an .asc file to be used for drawing random points}
}
\value{
output A data frame containing Monte Carlo replicates to be used in constructing 
        models for the background test.
}
\description{
This function generates replicates for the background test, as described in 
Warren et al. 2008
}
\examples{
reps.background.asc(infiles = "~/myfile.asc", outfile = "~/myreps.asc", reps=100)
}
\keyword{keywords}

