% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reps.rangebreak.line.R
\name{reps.rangebreak.line}
\alias{reps.rangebreak.line}
\title{reps.rangebreak.line, generates Monte Carlo replicates for the linear rangebreak
test of Glor and Warren 2011.}
\usage{
reps.rangebreak.line(infiles = x, outfile = FALSE, reps = 10,
  plotty = TRUE, background.raster = NA, verbose = FALSE)
}
\arguments{
\item{infiles}{One or more .csv files of species occurrence data}

\item{outfile}{Path to an output file.  If not provided, replicates will be returned 
in a data frame.}

\item{reps}{Number of replicates}

\item{plotty}{Controls plotting of replicates as they are constructed.}

\item{background.raster}{Raster to use as backrgound for plotting replicates}

\item{verbose}{Controls printing of diagnostic messages}
}
\value{
output A data frame containing replicates for the Monte Carlo test
}
\description{
This function applies linear breaks to the pooled distribution of two species.  Used 
in order to construct a null distribution of expected overlaps between species separated
by a linear boundary.
}
\examples{
reps.rangebreak.line("~/species_A_and_B.csv", reps=100)
}
\keyword{keywords}

